Refactored DFLCommunity into separate subcommunity
