+++ .gitignore
@@ -5,3 +5,4 @@ data/**
 __pycache__/
 *.py[cod]
 simulations/data/**
+simulations/dfl/data/**
